+++ PirateWeather-DotNetLib/PirateWeatherResponse.cs
@@ -20,8 +20,8 @@
         public int Time { get; set; }
         public string Summary { get; set; }
         public string Icon { get; set; }
-        public int NearestStormDistance { get; set; }
-        public int NearestStormBearing { get; set; }
+        public double NearestStormDistance { get; set; }
+        public double NearestStormBearing { get; set; }
         public double PrecipIntensity { get; set; }
         public double PrecipProbability { get; set; }
         public double PrecipIntensityError { get; set; }
@@ -33,7 +33,7 @@
         public double Pressure { get; set; }
         public double WindSpeed { get; set; }
         public double WindGust { get; set; }
-        public int WindBearing { get; set; }
+        public double WindBearing { get; set; }
         public double CloudCover { get; set; }
         public double UVIndex { get; set; }
         public double Visibility { get; set; }
@@ -80,7 +80,7 @@
         public double Pressure { get; set; }
         public double WindSpeed { get; set; }
         public double WindGust { get; set; }
-        public int WindBearing { get; set; }
+        public double WindBearing { get; set; }
         public double CloudCover { get; set; }
         public double UVIndex { get; set; }
         public double Visibility { get; set; }
@@ -122,7 +122,7 @@
         public double WindSpeed { get; set; }
         public double WindGust { get; set; }
         public int WindGustTime { get; set; }
-        public int WindBearing { get; set; }
+        public double WindBearing { get; set; }
         public double CloudCover { get; set; }
         public double UVIndex { get; set; }
         public int UVIndexTime { get; set; }
